- Fix `PaintBall` hooks was disabled - Fix wrong array count for `AresTEventExt::FindTechnoType` - `AresTEventExt::ResolveHouseParam` now return `nullptr` if the param value is -1 (need to be tested , is this gonna break stuffs or not) - disable most of `Hooks.EffectTargetCoordsFix.cpp` due to desync - Known Issues :      - Game will lockup during some survival map      - Some user still report random desyncs

+++ src/Ext/TEvent/Hooks.cpp
@@ -34,7 +34,7 @@ DEFINE_HOOK(0x7271F9, TEventClass_GetFlags, 0x5)
 	GET(TEventClass*, pThis, ESI);
 
 	int nEvent = static_cast<int>(pThis->EventKind);
-	if (nEvent >= PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= PhobosTriggerEvent::ShieldBroken)
+	if (nEvent >= (int)PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= (int)PhobosTriggerEvent::ShieldBroken)
 		eAttach |= TriggerAttachType::Logic; // LOGIC
 
 	R->EAX(eAttach);
@@ -48,7 +48,7 @@ DEFINE_HOOK(0x71F3FE, TEventClass_BuildINIEntry, 0x5)
 	GET(TEventClass*, pThis, ECX);
 
 	int nEvent = static_cast<int>(pThis->EventKind);
-	if (nEvent >= PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= PhobosTriggerEvent::ShieldBroken)
+	if (nEvent >= (int)PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= (int)PhobosTriggerEvent::ShieldBroken)
 		eNeedType = LogicNeedType::NumberNTech;
 
 	R->EAX(eNeedType);
@@ -61,7 +61,7 @@ DEFINE_HOOK(0x726577, TEventClass_Persistable, 0x7)
 	GET(TEventClass*, pThis, EDI);
 
 	int nEvent = static_cast<int>(pThis->EventKind);
-	if (nEvent >= PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= PhobosTriggerEvent::ShieldBroken)
+	if (nEvent >= (int)PhobosTriggerEvent::LocalVariableGreaterThan && nEvent <= (int)PhobosTriggerEvent::ShieldBroken)
 		R->AL(true);
 	else
 		R->AL(pThis->GetStateB());
